# Performance improvement for Guid.Equals using SSE