@jeffhandley 2021-07-23
> FWIW, I think that this is way too complex change for the benefit that it provides. I think it would be better to wait for the general Vector128 improvements.

Based on this comment and the lack of activity on this PR since, I'm going to go ahead and close it. Thanks for submitting the PR @billpoole-mi; it led to an interesting and valuable discussion here.